license => public domain
hilbert curve is now a "clean room" implementation based on the lindenmayer
system representation. the sloc grew a bit! on one hand it's a generic
lindenmayer system implementation, on the other hand the alternative curves
(peano and moore?) are more awkward to implement, and I'm ussure what value
they add.